+++ Documentation/Release-Notes/generate_release_notes.sh
@@ -3,7 +3,7 @@
 cd $(dirname $(realpath $0))
 
 major="`cd ../../ && cmake -P ./VERSION.cmake -L|grep 'MAXSCALE_VERSION_MAJOR'|sed 's/.*=//'`"
-minor="`cd ../../ && cmake -P ./VERSION.cmake -L|grep 'MAXSCALE_VERSION_MINOR'|sed 's/.*=//'`"
+minor="`cd ../../ && cmake -P ./VERSION.cmake -L|grep -v 'MAXSCALE_VERSION_MINOR_NUM' | grep 'MAXSCALE_VERSION_MINOR'|sed 's/.*=//'`"
 patch="`cd ../../ && cmake -P ./VERSION.cmake -L|grep 'MAXSCALE_VERSION_PATCH'|sed 's/.*=//'`"
 maturity="`cd ../../ && cmake -P ./VERSION.cmake -L|grep 'MAXSCALE_MATURITY'|sed 's/.*=//'`"